+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 - a way to specify nearest neighbor couplings in a :class:`~tenpy.models.lattice.Lattice`, 
   along with methods to count the number of nearest neighbors for sites in the bulk, and
   a way to plot them (:meth:`~tenpy.models.lattice.plot_coupling`)
+- :meth:`tenpy.networks.mpo.MPO.from_grids` to generate the MPO from a grid.
 
 Changed
 ^^^^^^^